requirements.txt
# Core dependencies
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0
PyYAML>=6.0

# NLP libraries
nltk>=3.6.5
transformers>=4.15.0
datasets>=1.18.0

# PyTorch (base requirement, install appropriate version for your system)
torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0

# Visualization libraries
matplotlib>=3.5.0
seaborn>=0.11.2
plotly>=5.5.0
wordcloud>=1.8.1

# Dimensionality reduction
umap-learn>=0.5.2

# For interactive visualizations
ipywidgets>=7.6.5

# For data handling and file operations
requests>=2.26.0
python-dateutil>=2.8.2
joblib>=1.1.0

# Advanced NLP tools
spacy>=3.2.0
sentencepiece>=0.1.96

# For working with different file formats
openpyxl>=3.0.9
xlrd>=2.0.1

# Utilities
pydantic>=1.9.0

# Testing tools (optional for development)
pytest>=6.2.5
pytest-cov>=2.12.1

# Packages for Model Training and Evaluation
tensorboard>=2.5.0
scikit-plot>=0.3.7
tensorboardX>=2.5.0

# Packages for Phase 5 - Explainability
shap>=0.41.0
lime>=0.2.0
captum>=0.5.0  # For additional interpretability

# Packages for Phase 5 - Deployment
onnx>=1.13.0
onnxruntime>=1.13.1
torchserve>=0.6.0
torch-model-archiver>=0.6.0
torch-workflow-archiver>=0.2.5

# Packages for Robust Error Handling
rich>=12.0.0  # Better logging and tracebacks
logging-color>=1.0.3  # Colored logging

# Additional requirements for Phase 5
protobuf>=3.20.0,<4.0.0  # Compatibility with ONNX
packaging>=21.0  # Necessary for version handling
sympy>=1.10.0  # Required by ONNX for symbolic operations
pillow>=9.0.0  # Required for image processing in some visualizations
colorama>=0.4.4  # For colored terminal outputs
tabulate>=0.8.9  # For table formatting in outputs